requirements.txt
python-telegram-bot==20.7
requests==2.31.0
asyncio
aiohttp==3.9.1
python-dotenv==1.0.0
base58==2.1.1
solders==0.21.0